Move new tests to correct location
